+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * NOTE: strict version checks on upload.
 
 ## `20211118t160200-all`
  * Bumped runtimeVersion to `2021.11.16`.